requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-JWT-Extended
Flask-Restx
Flask-Limiter
python-dotenv
Werkzeug
psycopg2-binary
pytest
requests